move submissions api tests to correct test class
